+++ app/updaters/updater.rb
@@ -1,0 +1,8 @@
+class Updater
+  using StringToModel
+  include DebugPrint
+  
+  def initialize(verbose: false)
+    @verbose = verbose
+  end
+end